Create License.txt
For what it's worth, going with GNU.org's recommendation for small programs. 

Small programs

It is not worth the trouble to use copyleft for most small programs. We use 300 lines as our benchmark: when a software package's source code is shorter than that, the benefits provided by copyleft are usually too small to justify the inconvenience of making sure a copy of the license always accompanies the software.

For those programs, we recommend the Apache License 2.0. This is a pushover (non-copyleft) software license that has terms to prevent contributors and distributors from suing for patent infringement. This doesn't make the software immune to threats from patents (a software license can't do that), but it does prevent patent holders from setting up a “bait and switch” where they release the software under free terms then require recipients to agree to nonfree terms in a patent license.

Among the lax pushover licenses, Apache 2.0 is best; so if you are going to use a lax pushover license, whatever the reason, we recommend using that one.